Implement Stefano and Chuntai comments
